requirements.txt
django==3.0
django_reverse_admin
django-phonenumber-field
phonenumbers
